DDOL object creation
gameMasterObj created on devSplash(since it is the first scene in the current build order)
Obj has a script attached that creates a list for items and loads in all current made items. This is NOT* the inventory. This is for item drops and shop purposes. The inventory will be held in the User Object.